\name{rbcl}
\docType{data}
\alias{rbcl}
\title{Green Plant rbcL}
\description{This dataset has been used by Xie et al. (2011) and has 10 species of green plant.  The DNA sequences are from the chloroplast-encoded large subunit of the RuBisCO gene (rbcL).
}
\usage{data(rbcl)}
\format{\code{phyDat} object.}
\source{https://academic.oup.com/sysbio}
\references{
Xie, W., Lewis, P. O., Fan, Y., Kuo, L., Chen, M.-H., 2011. Improving marginal
likelihood estimation for Bayesian phylogenetic model selection.  \emph{Systematic Biology} \bold{60}(2), 150--160.
}
\keyword{datasets}
